Extract HasZipModels super class
This can be shared between the place recognizer and the file-based
sentiment detector.
